requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
numpy==2.2.2
pandas==2.2.3
pandas_ta==0.3.14b0
pyarrow==19.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
tqdm==4.67.1
tzdata==2025.1
urllib3==2.3.0
